--- a/arm-wt-22k/host_src/eas.cmake
+++ b/arm-wt-22k/host_src/eas.cmake
@@ -46,6 +46,8 @@ extern "C" {
 #define MAKE_LIB_VERSION(a,b,c,d) (((((((EAS_U32) a <<8) | (EAS_U32) b) << 8) | (EAS_U32) c) << 8) | (EAS_U32) d)
 #define LIB_VERSION MAKE_LIB_VERSION(@PROJECT_VERSION_MAJOR@,@PROJECT_VERSION_MINOR@,@PROJECT_VERSION_PATCH@,@PROJECT_VERSION_TWEAK@)
 
+#cmakedefine NEW_HOST_WRAPPER
+
 typedef struct
 {
     EAS_U32     libVersion;