After profiling about DLS loading, I found EAS's HW was so slow about file operations. It was designed for twenty years ago.
So I reimplemented it, using CRT and some OS-specific APIs. The old one is preserved as optional.

It works on Linux and Windows now, *blazing fast* (compared than before).