@rsp4jack 2024-11-02
The CRT file operations are only used when file->readAt is NULL.