@rsp4jack 2024-11-03
Benchmark (not authoritative, for information only):

OmegaGMGS2.dls, ~266 MiB, with 484 insts, 15947 regions, 15944 arts and 2255 waves
Old HW:
```
./sonivoxrender -d ~/OmegaGMGS2.dls  23.50s user 22.76s system 99% cpu 46.503 total
```
New HW:
```
./sonivoxrender -d ~/OmegaGMGS2.dls  0.41s user 0.20s system 99% cpu 0.612 total
```