@rsp4jack 2024-11-09
> Where can I find OmegaGMGS2.dls? I can only find the sf2 version of it.

It is converted from the sf2 version, using Extreme Sample Converter.
BTW, the eas_mdls limits are increased to make it load.