@pedrolcl 2024-11-09
> > Where can I find OmegaGMGS2.dls? I can only find the sf2 version of it.
> 
> It is converted from the sf2 version, using Extreme Sample Converter.

I've tested [Airfont 340.dls](http://www.ronimusic.com/sf2/Airfont_340.dls), ~80 MiB [from this page](http://www.ronimusic.com/smp_ios_dls_files.htm) and the benefit is evident. 

Please take a look to my modifications. I want to do more checks before a new release.