@rsp4jack 2024-11-10
> Please take a look to my modifications. I want to do more checks before a new release.

EAS's duplicated handle has its own position indicator, but as I know, dup()-ed file descriptors share the same I/O object (so file position is same).

Use dup() to implement DupHandle will not work on Linux (as I tested it), but according to CI it works on FreeBSD.