@pedrolcl 2024-11-11
> according to CI it works on FreeBSD and MacOS. Confused about how it works

CI is only proving that the code is compiling, and the unit tests are passing. But unit tests are only testing MIDI file processing without DLS files. Other manual tests with the sonivoxrender program show that FreeBSD reads correctly all DLS files, but macOS doesn't. What is needed:
- Unit tests that include DLS. The main problem is having a free and reasonable sized DLS to be included in the repository for testing.
- I am not happy with the special Linux implementation. There should be a single DupHandle implementation for all POSIX platforms.

The alternative right now is to enable at configuration time the new host wrapper for some platforms and the old wrapper for others. This doesn't feel right for a release.