@rsp4jack 2024-11-17
I found that even passing ReadAt to the new HW, it is still much faster than the old one:

```
./sonivoxrender -d ~/OmegaGMGS2.dls  3.19s user 1.80s system 99% cpu 5.022 total
```

> The alternative right now is to enable at configuration time the new host wrapper for some platforms and the old wrapper for others. This doesn't feel right for a release.

For these platforms, fallback to internal ReadAt in the new HW is okay. I will implement it in the next PR.